STL file visualisation now implemented with display lists this is WAAAAAY faster ^^
The rotation around all axes is only added in this commit to show that this works very fast (for comparison set draw_modes=DrawingModes.triangles within
the parameter list of the StlModelViews __init__ function and see how "fast" the rotation works ....)

As a note: The initialization of this display lists absolutely needs to be done after the opengl initialization is done.
So an easy solution for this (as we have done here) is to add an initialized flag and check within the draw function if the geometry has already been initialized and disable the flag after initialization.
This way the display list is build during the first call of the objects draw function.